backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.33.1",
    "uvicorn>=0.46.0",
]
